fix: added global type
